cloud-run-functions/bot-response/requirements.txt
google-cloud-bigquery
google-cloud
google-genai
python-dotenv


